Changed barcode column type to bigInt to store ean13 barcodes

+++ database/migrations/2025_01_15_064739_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('sku')->unique();
             $table->string('name')->nullable();
             $table->integer('barcode')->nullable();
-            $table->integer('quantity')->nullable();
+            $table->unsignedBigInteger('quantity')->nullable();
             $table->timestamps();
         });
     }
